requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.0,<4.0.0
djangorestframework-simplejwt>=5.2.2
psycopg2-binary>=2.9.6
drf-yasg>=1.21.5
gunicorn>=20.1.0
